+++ src/main/java/com/cleevio/vexl/common/integration/coingecko/dto/response/MarketData.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Data
@@ -14,25 +15,25 @@ public class MarketData {
     private CurrentPrice currentPrice;
 
     @JsonProperty("price_change_percentage_24h")
-    private Double priceChangePercentage24h;
+    private BigDecimal priceChangePercentage24h;
 
     @JsonProperty("price_change_percentage_7d")
-    private Double priceChangePercentage7d;
+    private BigDecimal priceChangePercentage7d;
 
     @JsonProperty("price_change_percentage_14d")
-    private Double priceChangePercentage14d;
+    private BigDecimal priceChangePercentage14d;
 
     @JsonProperty("price_change_percentage_30d")
-    private Double priceChangePercentage30d;
+    private BigDecimal priceChangePercentage30d;
 
     @JsonProperty("price_change_percentage_60d")
-    private Double priceChangePercentage60d;
+    private BigDecimal priceChangePercentage60d;
 
     @JsonProperty("price_change_percentage_200d")
-    private Double priceChangePercentage200d;
+    private BigDecimal priceChangePercentage200d;
 
     @JsonProperty("price_change_percentage_1y")
-    private Double priceChangePercentage1y;
+    private BigDecimal priceChangePercentage1y;
 
     @JsonProperty("last_updated")
     private ZonedDateTime lastUpdated;
